Add proxy handlers for external routing services
This adds three handlers for OSRM, GraphHopper and Mapbox, which
essentially proxy the requests made to the routing APIs and inject the
required API keys if needed.

The reason for doing this is to offer protection for both the user and
administrator, because first of all, external services can then no
longer track usage down to individual users but essentially get requests
from one source only. Second, the administrator no longer has his/her
API keys exposed to all of the users.

Note that this change only adds the routes and handlers for proxying and
it's not yet integrated into the frontend.

Unfortunately, when passing along and injecting API keys into the query
string we need to parse query string arguments by ourselves because
GraphHopper uses the "point" query string argument multiple times and
PHP by default treats them as unique keys.

Signed-off-by: aszlig <aszlig@nix.build>