script: Switch to new proxying routes
This switches the main JavaScript file to use our shiny new proxying
handlers instead of directly requesting the API via the user's browser.

Since the API keys are solely handled by the backend, we now just pass
null for Mapbox and GraphHopper instead of the API key, because it gets
replaced before the proxy handler submits the request.

In addition to the optionsValues we had so far, we now also have a new
availableFeatures object, which only contains the settings necessary to
know whether a service is available or not rather than its details, like
eg. API keys and URLs.

Signed-off-by: aszlig <aszlig@nix.build>