proxy: Replace usage of cURL by OCP\Http\Client
When looking for other uses of HTTP client in the Nextcloud server, I
found a few usages of cURL and thought this would actually be the best
practice.

However, it turns out that there is actually an HTTP client
implementation based on Guzzle, so we can completely throw away our
custom ProxyResponse.

Additionally, I also checked the response format of the routing APIs we
use and all of them are using JSON, so instead of using something like
DataDisplayResponse (which adds a Content-Disposition header) or yet
again implement our custom Response, I just changed this to use
JSONResponse instead.

What I kept however is the setting of the User-Agent header, because
some of the API providers tend to have a policy that require you to
properly set it, for example for the OSRM demo server their policy[1]
states:

> General rules
>
>  ...
>
>  * Valid User-Agent identifying application. Faking another app's
>    User-Agent WILL get you blocked.

[1]: https://github.com/Project-OSRM/osrm-backend/wiki/Api-usage-policy

Signed-off-by: aszlig <aszlig@nix.build>