Fix Add Device Point error when optional value is empty.
This happens whith PhoneTrack App. I found logs of it sending requests with query params as follows: ?lat=1.1&lng=2.2&timestamp=1111111111&user_agent=testDevice&altitude=&battery=95.0&accuracy=36.0
In the above example the altitude is present with a value of empty string (""). This value was passed to the DB sevice, which would fail due to the value not being a number.

Signed-off-by: Alessio Linares <alessio@alessio.cc>