Rename "features" variable to "routingProviders"
This was a suggestion from @eneiluj:

> Did you call it "features" because you think ahead and predict it can
> be used for other types of features? I wonder if it could be named
> more explicitly like "routingProviders" and leave other types of
> features in another key to make it clearer. I mean "routing" is a
> feature but "mapbox" is not...

I originally used "features", because it could be used to expose more
than just routing service information. In hindsight however it's a
better idea to call it "routingProviders" (as suggested), so it's more
clear when you read the code. Should we add something like map tile
services, we can still add another variable, eg. "tileProviders".

Signed-off-by: aszlig <aszlig@nix.build>