PageController: allow tile.openstreetmap.org in CSP
So far, we only allowed subdomains, but when I rolled out the other
changes, this was missing from having a working map with OSM tiles.

This is probably not related to the other changes (it worked fine a few
days ago, perhaps something was just changed within OSM), but I figured
it's still worth fixing.

Signed-off-by: Maximilian Bosch <maximilian@mbosch.me>